Potentially revamp to PWM driver instead of SPI

+++ TestSoftware/SegmentDriver/SegmentDriver/SPI.h
@@ -2,5 +2,5 @@
 #include <stdbool.h>
 #include "stm32f0xx.h"
 void InitSPI();
-void SendSPIDMA(uint8_t* data, uint8_t bytes);
+void SendSPIDMA(uint8_t* data, uint16_t bytes);
 bool isDMAFinished();
